extractParams check the ParametersBag for query
The existing extractParams assumed the query would be directly on the body of the request. This is a problem if you need to send multi-part form data like in the instance of uploading files. There needs to be a way to extract the query from the POST parameters bag in this case.