Entity reference: Scope direct queries
Previously, all the queries where exposed straight on the entity type, not
taking into account the bundles on which the field actually appears.

This changes it to keep exposing the base fields on the entity type interface,
but expose the configurable entity reference fields only on the bundles they're
attached to.