feat(graphql_core): redirect entities also resolve with query parameters
By using the full path (including query parameters) to resolve to a
redirect entity before, it didn't work when your query params weren't
corresponding with the ones in the redirect entity values. This is not
redirect module's core behaviour (as it works regardless of query
parameters). This patch resolves this by resolving the redirect entity
by path, and also implement the "Retain query string" setting and merge
the query strings when enabled.