Allow non-existent type reference while importing config
When creating a plugin that references a type that does not exist, the graphql module generates a fatal error.

This makes sense, except when importing config. When config is imported, first all modules are installed before other config (content types, fields) are added. This makes it impossible to do a 'clean install' when a plugin references a type that is automatically generated from Drupal.

This merge request ignores missing type associations when `\Drupal::isConfigSyncing()` returns `TRUE`.