Don't create new Deferred in returnFinally
A new deferred object used to be needed because `then` was called on and
returned the wrapped promise of the `Deferred` object, which confused
calling code.

However, as of v14.0.0 of the graphql-php library the Deferred object
extends the promise itself so `then` properly creates a new `Deferred`
instance which will still work well with `instanceof` checks.