Update debug configuration for v14.0.0
The `setDebug` method has been changed to `setDebugFlag`. In order to
keep the user interface simple for people creating Drupal Server
entities, this commit still keeps the binary "on/off" behaviour for
debugging, using the new function.

The only difference with the previous implementation is that `setDebug`
in GraphQL would hide the flag but default to `1` (casting `(int)TRUE`).
We now use the `2` value flag, which includes a backtrace with the
error. This is more useful when debugging problems.