Accept SyncPromise where Deferred was accepted
This commit updates code that checks for whether a promise needs to be
resolved to accept `SyncPromise` instead of `Deferred`.

Calling `then` on a `Deferred` instance will return a new `SyncPromise`
instance that is the next promise in the chain. This is caused by the
`SyncPromise` implementation calling `self`.

This change allows a chained promise to be easily usable in resolvers.
For example the following is now valid.

```
// in someclass
public function resolve() {
  return $this->someMethodReturningDeffered()->then(function ($x) {
   // modify $x
   return $x;
  })->then(function ($y) {
   // modify $y
   return $y;
  });
}
```

The above code would previously lead to issues because a `SyncPromise`
would be treated as a value instead of a deferred result. Before this
change all values would have to be put in new `Deferred` instances which
makes using promises a lot more difficult.