Rewrite fromPath to use PropertyPath data producer
The `fromPath` method is a really useful shortcut for
a commonly used pattern. However, it uses the `Path`
resolver which is marked for removal.

An additional issue is encountered when trying to use
multiple `fromPath` resolvers for fields on the same
piece of data. These will work individually but will cause
resolver errors when used in a query at the same time.
This is probably caused by a slight difference in handling
of value between the PropertyPath DataProducer and the
Path Resolver.

Rewriting the `fromPath` property in terms of the
`PropertyPath` data resolver retains the useful
shortcut but ensures it's functionally equivalent to the
long form.