Change `debug` to `debug_flag` in server configuration
The underlying library now allows to set the debugging configuration
using a set of enum values forming a bitmask.

The Server form is altered to show the individual options so that
developers can select them. Additionally the use of `debug` in the code
is changed to `debug_flag` to more clearly show how everything ties
together.

An update hook is included to ensure that if people are currently
debugging they retain most of the information they expect.