fix(ui): Fix PHP warning for non-variables passed by reference (#1080)
* Fix notice about variables passed by reference

`reset` only takes variables so we can not pass the result of
`array_keys` directly but have to use an intermediate variable.

* Merge variable fallback expressions

`array_key_exists('schema', $input) ? $input['schema'] : NULL` is equal
to `$input['schema'] ?? NULL` on PHP versions that support the
null-coalescing operator. Since we already use this elsewhere we already
require such PHP versions.

This allows us to substitute the simplified one-use variable leading to
`$input['schema'] ?? NULL ?? $server->get('schema')`. The expression `??
NULL ??` can be safely transformed to `??` to result in the simplified
expression in this commit.

Co-authored-by: Klaus Purer <klaus.purer@protonmail.ch>