Use test module with data type for path resolver test
It was impossible to get the path resolver working with a mocked
TypedDataManager instance since the logic around data producers also
uses it to get the inputs for the PropertyPath DataProducer.

To work around this a test module is created that declares a nestable
DataType that can be used to create unmocked test data.

This changes the test data for the test but not the query or assertions
under test.