Implement interface resolver resolution
GraphQL supports interfaces on types and on other interfaces. This
change makes it possible to define a resolver for an interface type,
this can then be used by any type that extends the interface. This makes
it easier to implement patterns like connections and edges.

A resolver defined for a concrete type will take precedence.

This bumps the minimum version of graphql-php to 14.5.0 as that is where
the `ImplementingType` interface is added which is used to figure out if
a type can have any ancestors.